correcciones algortimos

+++ kernel/src/pcb.h
@@ -25,8 +25,8 @@ typedef struct
     t_estado_proceso estado_actual;
     t_temporal* temporal_estado;
     int tamanio;
-    double estimacion_rafaga;
-    double rafaga_real_anterior;
+    int estimacion_rafaga;
+    int rafaga_real_anterior;
     t_temporal* temporal_blocked;
 } t_pcb;
